using the same dotdev config used for other applications
