+++ wedlocks/package.json
@@ -10,6 +10,7 @@
     "@babel/preset-env": "^7.3.4",
     "aws-sdk": "^2.423.0",
     "babel-loader": "^8.0.5",
+    "dotenv": "^8.0.0",
     "netlify-cli": "^2.8.3",
     "nodemailer": "^6.1.0",
     "nodemon": "^1.19.0",
@@ -28,8 +29,8 @@
   "devDependencies": {},
   "scripts": {
     "build": "webpack",
-    "dev": "nodemon src/server.js --exec babel-node",
-    "test": "jest",
+    "dev": "UNLOCK_ENV=dev nodemon src/server.js --exec babel-node",
+    "test": "UNLOCK_ENV=test jest",
     "lint": "eslint --ext .ts,.js src/",
     "ci": "npm run lint && npm test",
     "deploy-netlify": "./scripts/deploy-netlify.sh"
